Wait for wget jobs to complete
Have the `fetch` script wait for all wget jobs to complete instead of
forking them all immediately. This provides a mechanism to do some kind
of "post processing" once the archive is generated.

Note that in order to generate the list of PIDs we have to pipe the
output of awk to a temp file and then read the temp file in the while
loop, instead of simply piping awk into the while loop directly. This is
due to the fact that variables modified within a pipeline do not have
their values modified. See SC2031 [1] for more information.

[1]: https://github.com/koalaman/shellcheck/wiki/SC2031